Fix multi-parameter calls (temporary fix)
Calls with multiple parameters were failing because the appropriate mnemonics were not being used for the appropriate data type sizes (MOVQ was used everywhere). I have temporarily fixed this by ensuring every variable uses up 8 bytes of space, but the proper fix is to change the mnemonics in accordance with the data type.