Implement subscript expressions
This commit adds support for subscript expressions, aka pointer dereferencement. It adds a new kind of of Operand: `IndirectionOperand`. It has a single attribute: `attribute` that is an other operand of type `Int64DataType`. It represents a value in memory.

A new type can be passed to the assembly now: `PointerDataType`. It can occur as the left operand and output of a binary expression (then the right side is an `Int64`) or in CopyOp / CastOp.

You can also receive an ArrayDataType as the right member of a CastOp.

Signed-off-by: Charles Samborski <demurgos@demurgos.net>