Fix branching for nested control flow statements
The branching logic was applied after the IR of the body of the control
flow statements was built. The issue is that when the body
contained control flow statements itself (nested), it created new basic
blocks and _updated the reference_. It led to the top branches
to point to point to _the last block_ of the body.

Signed-off-by: Charles Samborski <demurgos@demurgos.net>